requirements.txt
aiohttp
asgiref
Django~=5.2.4
aiogram~=3.21.0
dotenv==0.9.9
django-unfold==1.0.0